+++ src/components/page-layout/HomeLayout/components/PostList/GiverPostList/GiverPostList.module.scss
@@ -15,3 +15,9 @@
   grid-template-columns: repeat(4, 1fr);
   gap: 3rem;
 }
+
+.heart {
+  position: absolute;
+  top: 1rem;
+  right: 2rem;
+}
