+++ src/components/page-layout/HomeLayout/components/PostList/TakerPostList/TakerPostList.tsx
@@ -3,7 +3,8 @@ import classNames from "classnames/bind";
 
 import Link from "next/link";
 
-import { Post, PostHeart } from "@/components/common/Post/Post";
+import { Post } from "@/components/common/Post/Post";
+import { PostHeart } from "@/components/common/Post/PostHeart/PostHeart";
 import styles from "@/components/page-layout/HomeLayout/components/PostList/TakerPostList/TakerPostList.module.scss";
 import { ROUTE } from "@/constants/route";
 import Plus from "@/icons/plus.svg";
@@ -30,7 +31,7 @@ export default function TakerPostList() {
       <div className={cn("postListBox")}>
         {data.content.map((post: PostType) => (
           <Post data={post} key={post.id}>
-            <PostHeart queryKey={["mainHelpMePostList"]} id={post.id} isLiked={post.isLiked} />
+            <PostHeart queryKey={["mainHelpMePostList"]} id={post.id} isLiked={post.isLiked} style={cn("heart")} />
           </Post>
         ))}
       </div>
