+++ src/components/page-layout/helpMeLayout/components/helpMeLayout.module.scss
@@ -51,3 +51,9 @@
 .button:hover .arrow {
   fill: $taker;
 }
+
+.heart {
+  position: absolute;
+  top: 1rem;
+  right: 2rem;
+}
