+++ src/components/page-layout/helpYouDetailLayout/components/helpYouDetailLayout.module.scss
@@ -31,7 +31,7 @@
   position: relative;
 }
 
-.likeBtn,
+.heart,
 .kebabBtn,
 .sirenBtn {
   @include flexSort(column, null, center, 0.2rem);
