+++ src/components/page-layout/helpYouLayout/components/helpYouLayout.module.scss
@@ -51,3 +51,9 @@
 .button:hover .arrow {
   fill: $giver;
 }
+
+.heart {
+  position: absolute;
+  top: 1rem;
+  right: 2rem;
+}
