+++ src/components/page-layout/helpYouLayout/components/helpYouLayout.tsx
@@ -9,7 +9,8 @@ import { useRouter } from "next/router";
 import Filter from "@/components/common/Filter/Filter";
 import getPaginationItems from "@/components/common/Pagenation/apis/getHelpMeList";
 import Pagination from "@/components/common/Pagenation/Pagenation";
-import { Post, PostHeart } from "@/components/common/Post/Post";
+import { Post } from "@/components/common/Post/Post";
+import { PostHeart } from "@/components/common/Post/PostHeart/PostHeart";
 import styles from "@/components/page-layout/helpYouLayout/components/helpYouLayout.module.scss";
 import { ROUTE } from "@/constants/route";
 import RegisterArrow from "@/icons/send_arrow.svg";
@@ -130,6 +131,7 @@ export default function HelpYouLayout() {
                 queryKey={["helpYouPostList", `${page}`, disabilityType, assistanceType, postStatus]}
                 id={post.id}
                 isLiked={post.isLiked}
+                style={cn("heart")}
               />
             </Post>
           ))}
